Clean-up of warnings in usr/ builds.
o Added entry to latest changelog to close the apex-env --force bug.
o Added -Wall to usr/ builds so we can see warnings.
o Fixed warnings as a coupld of bonafide errors in apex-env.

+++ usr/link.h
@@ -85,7 +85,7 @@ protected:
 
   int fhEnvChar;		// File handle for NOR-wise writes
   int fhEnvBlock;		// File handle for erasing
-  int ibEnv;			// Index of environment data in file handles
+  ssize_t ibEnv;                // Index of environment data in file handles
 
   EntryMap* entries;		// Entries found in flash
   int idNext;			// Next available ID for flash environment
